Extract reusable hand or pouch placement
